+++ backend/src/test/java/com/ssafy/s309/domain/user/service/UserServiceTest.java
@@ -33,7 +33,7 @@ class UserServiceTest {
   @Mock private WardGuardianRepository wardGuardianRepository;
   @InjectMocks private UserService userService;
 
-  private static final Long USER_ID = 1L;
+  private static final Integer USER_ID = 1;
 
   private User ward;
 
@@ -100,15 +100,15 @@ class UserServiceTest {
   void 보호자_목록_조회_priority_순서_반환() {
     User g1 = User.builder().email("g1@example.com").build();
     User g2 = User.builder().email("g2@example.com").build();
-    ReflectionTestUtils.setField(g1, "id", 2L);
-    ReflectionTestUtils.setField(g2, "id", 3L);
+    ReflectionTestUtils.setField(g1, "id", 2);
+    ReflectionTestUtils.setField(g2, "id", 3);
 
     WardGuardian wg1 =
         WardGuardian.builder().ward(ward).guardian(g1).relation("부모").priority(0).build();
     WardGuardian wg2 =
         WardGuardian.builder().ward(ward).guardian(g2).relation("배우자").priority(1).build();
-    ReflectionTestUtils.setField(wg1, "id", 10L);
-    ReflectionTestUtils.setField(wg2, "id", 11L);
+    ReflectionTestUtils.setField(wg1, "id", 10);
+    ReflectionTestUtils.setField(wg2, "id", 11);
 
     given(userRepository.findById(USER_ID)).willReturn(Optional.of(ward));
     given(wardGuardianRepository.findAllByWard_IdOrderByPriorityAsc(USER_ID))
@@ -125,36 +125,36 @@ class UserServiceTest {
   @Test
   void 보호자_등록_priority_자동_할당() {
     User guardian = User.builder().email("guardian@example.com").build();
-    ReflectionTestUtils.setField(guardian, "id", 2L);
+    ReflectionTestUtils.setField(guardian, "id", 2);
 
     WardGuardian saved =
         WardGuardian.builder().ward(ward).guardian(guardian).relation("부모").priority(2).build();
-    ReflectionTestUtils.setField(saved, "id", 10L);
+    ReflectionTestUtils.setField(saved, "id", 10);
 
     given(userRepository.findById(USER_ID)).willReturn(Optional.of(ward));
-    given(userRepository.findById(2L)).willReturn(Optional.of(guardian));
+    given(userRepository.findById(2)).willReturn(Optional.of(guardian));
     given(wardGuardianRepository.countByWard_Id(USER_ID)).willReturn(2);
     given(wardGuardianRepository.save(any(WardGuardian.class))).willReturn(saved);
 
-    GuardianResponse response = userService.createGuardian(USER_ID, new GuardianRequest(2L, "부모"));
+    GuardianResponse response = userService.createGuardian(USER_ID, new GuardianRequest(2, "부모"));
 
-    assertThat(response.guardianId()).isEqualTo(2L);
+    assertThat(response.guardianId()).isEqualTo(2);
     assertThat(response.priority()).isEqualTo(2);
   }
 
   @Test
   void 보호자_수정_성공() {
     User guardian = User.builder().email("guardian@example.com").build();
-    ReflectionTestUtils.setField(guardian, "id", 2L);
+    ReflectionTestUtils.setField(guardian, "id", 2);
 
     WardGuardian wg =
         WardGuardian.builder().ward(ward).guardian(guardian).relation("부모").priority(0).build();
-    ReflectionTestUtils.setField(wg, "id", 10L);
+    ReflectionTestUtils.setField(wg, "id", 10);
 
-    given(wardGuardianRepository.findById(10L)).willReturn(Optional.of(wg));
+    given(wardGuardianRepository.findById(10)).willReturn(Optional.of(wg));
 
     GuardianResponse response =
-        userService.updateGuardian(USER_ID, 10L, new GuardianRequest(2L, "가족"));
+        userService.updateGuardian(USER_ID, 10, new GuardianRequest(2, "가족"));
 
     assertThat(response.relation()).isEqualTo("가족");
   }
@@ -162,18 +162,17 @@ class UserServiceTest {
   @Test
   void 다른_유저의_보호자_수정_예외() {
     User otherWard = User.builder().email("other@example.com").build();
-    ReflectionTestUtils.setField(otherWard, "id", 99L);
+    ReflectionTestUtils.setField(otherWard, "id", 99);
 
     User guardian = User.builder().email("guardian@example.com").build();
-    ReflectionTestUtils.setField(guardian, "id", 2L);
+    ReflectionTestUtils.setField(guardian, "id", 2);
 
     WardGuardian wg = WardGuardian.builder().ward(otherWard).guardian(guardian).priority(0).build();
-    ReflectionTestUtils.setField(wg, "id", 10L);
+    ReflectionTestUtils.setField(wg, "id", 10);
 
-    given(wardGuardianRepository.findById(10L)).willReturn(Optional.of(wg));
+    given(wardGuardianRepository.findById(10)).willReturn(Optional.of(wg));
 
-    assertThatThrownBy(
-            () -> userService.updateGuardian(USER_ID, 10L, new GuardianRequest(2L, null)))
+    assertThatThrownBy(() -> userService.updateGuardian(USER_ID, 10, new GuardianRequest(2, null)))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("해당 유저의 보호자가 아닙니다");
   }
@@ -183,20 +182,20 @@ class UserServiceTest {
     User g1 = User.builder().email("g1@example.com").build();
     User g2 = User.builder().email("g2@example.com").build();
     User g3 = User.builder().email("g3@example.com").build();
-    ReflectionTestUtils.setField(g1, "id", 2L);
-    ReflectionTestUtils.setField(g2, "id", 3L);
-    ReflectionTestUtils.setField(g3, "id", 4L);
+    ReflectionTestUtils.setField(g1, "id", 2);
+    ReflectionTestUtils.setField(g2, "id", 3);
+    ReflectionTestUtils.setField(g3, "id", 4);
 
     WardGuardian wg0 = WardGuardian.builder().ward(ward).guardian(g1).priority(0).build();
     WardGuardian wg1 = WardGuardian.builder().ward(ward).guardian(g2).priority(1).build();
     WardGuardian wg2 = WardGuardian.builder().ward(ward).guardian(g3).priority(2).build();
-    ReflectionTestUtils.setField(wg1, "id", 11L);
+    ReflectionTestUtils.setField(wg1, "id", 11);
 
-    given(wardGuardianRepository.findById(11L)).willReturn(Optional.of(wg1));
+    given(wardGuardianRepository.findById(11)).willReturn(Optional.of(wg1));
     given(wardGuardianRepository.findAllByWard_IdOrderByPriorityAsc(USER_ID))
         .willReturn(List.of(wg0, wg2));
 
-    userService.deleteGuardian(USER_ID, 11L);
+    userService.deleteGuardian(USER_ID, 11);
 
     verify(wardGuardianRepository).delete(wg1);
     assertThat(wg2.getPriority()).isEqualTo(1);
